Refactoring module and keeping it off the reactor build
