+++ kie-dmn-feel/src/test/resources/org/kie/dmn/feel/lang/examples/TEST.feel
@@ -1,0 +1,5 @@
+{
+   var1 : "Hello World" ,
+   
+   2  
+}
