+++ kie-dmn-model/src/main/java/org/kie/dmn/feel/model/v1_1/DMNModelInstrumentedBase.java
@@ -61,6 +61,9 @@ public abstract class DMNModelInstrumentedBase {
         this.parent = parent;
     }
 
+    /*
+     * children element references are populated during deserialization, enabling fast access for Validation.
+     */
     public java.util.List<DMNModelInstrumentedBase> getChildren() {
         return Collections.unmodifiableList(children);
     }
