+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/BKM_MISSING_VAR.dmn
