+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/DTABLE_MULTIPLEOUTPUT_WRONG_OUTPUT.dmn
