+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/DTABLE_SINGLEOUTPUT_WRONG_OUTPUT.dmn
