+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/ELEMREF_NOHASH.dmn
