+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/ITEMCOMP_DUPLICATED.dmn
