+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/RELATION_DUP_COLUMN.dmn
