+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/REQAUTH_NOT_KNOWLEDGESOURCE.dmn
