+++ kie-dmn-validation/src/test/resources/org/kie/dmn/validation/TYPEREF_NO_FEEL_TYPE.dmn
