+++ pom.xml
@@ -45,7 +45,6 @@
     <module>kie-dmn-feel</module>
     <module>kie-dmn-backend</module>
     <module>kie-dmn-core</module>
-    <module>kie-dmn-validator</module>
   </modules>
 
   <dependencyManagement>
